+++ airflow/.gitignore
@@ -1,2 +1,3 @@
 .env_dag
 .env
+.DS_Store
