+++ airflow/Dockerfile
@@ -1,4 +1,4 @@
-FROM apache/airflow:2.10.2
+FROM apache/airflow:2.8.1
 
 COPY requirements.txt .
 RUN pip install --upgrade pip
