+++ airflow/postgres_data/postmaster.pid
@@ -1,8 +1,8 @@
 1
 /var/lib/postgresql/data
-1728998391
+1729017069
 5432
 /var/run/postgresql
 *
-       29         0
+        7         0
 ready
